move serialization code to chain/
