+++ gen/gen.go
@@ -134,7 +134,7 @@ func MakeMethods(jenFile *jen.File, details GenDetails) {
 		).Params(
 			jen.Id("*Message"),
 		).Block(
-			jen.Id("ser").Op(":=").Id("state.MustSerialize").Call(jen.Id("&params")),
+			jen.Id("ser").Op(":=").Id("MustSerialize").Call(jen.Id("&params")),
 			jen.Return(jen.Id("mp.Build").Params(jen.Id("to"), jen.Id("from"), jen.Id(fmt.Sprintf("builtin_spec.Methods%s", d.methodPrefix)), jen.Id("ser"), jen.Id("opts..."))),
 		)
 	}
