+++ pkg/chain/serialization.go
@@ -1,4 +1,4 @@
-package state
+package chain
 
 import (
 	"bytes"
